Pass down download_models
Also tweak a related portion of the README to make the parameter
name match and update the github_install commands to refer to
the specific repo.

Note: I haven't been fully able to run the model due to running
into issues with downloading files, including a potentially missing
model file (for the "encoder" portion).